Core

Tip

We suggest that you read through this section if you wish to see how the the library works. To change the model constants (in case you have a different system), see the respective src/{module}/constants/constants.jl file.

CaseStudy

Modules

CaseStudy.CaseStudy — Module

CaseStudy core, allows access to the following submodules:

SenthilModel

AshwiniModel

Types

Functions

Constants

CaseStudy.SenthilModel

Modules

CaseStudy.SenthilModel — Module

Implementation of reduced order model for a lithium ion cell with uniform reaction rate approximation.

For model details refer here.

Structs

CaseStudy.SenthilModel.SolidConcentration — Type

Volume averaged solid concentration

Fields

c₁ₙ: Solid phase concentration in active material spheres in negative electrode.

c₁ₚ: Solid phase concentration in active material spheres in positive electrode.

CaseStudy.SenthilModel.InterfacialConc — Type

Interfacial electrolyte concentration

Fields

c₂ᵢₙ: Interfacial electrolyte concentration in negative electrode.

c₂ᵢₚ: Interfacial electrolyte concentration in positive electrode.

CaseStudy.SenthilModel.InterfacialFlux — Type

Interfacial flux

Fields

q₂ᵢₙ: Interfacial flux in negative electrode.

q₂ᵢₚ: Interfacial flux in positive electrode.

CaseStudy.SenthilModel.SolidRadialConcentration — Type

Radially averaged solid concentration

Fields

c₁ₙᵣ: Solid phase concentration in active material spheres in negative electrode.

c₁ₚᵣ: Solid phase concentration in active material spheres in positive electrode.

Functions

CaseStudy.SenthilModel.SOCₙ — Method

State of Charge in negative electrode

CaseStudy.SenthilModel.SOCₚ — Method

State of Charge in positive electrode

CaseStudy.SenthilModel.Uₙ — Method

Open circuit potential in negative electrode

CaseStudy.SenthilModel.Uₚ — Method

Open circuit potential in positive electrode

CaseStudy.SenthilModel.V — Method

Cell voltage across the battery

CaseStudy.SenthilModel.c₂ — Method

Overall electrolyte concentration function

CaseStudy.SenthilModel.c₂̄ₙ — Method

Average electrolyte concentration in negative electrode

CaseStudy.SenthilModel.c₂̄ₚ — Method

Average electrolyte concentration in positive electrode

CaseStudy.SenthilModel.c₂̄ₛ — Method

Average electrolyte concentration in separator

CaseStudy.SenthilModel.c₂ₙ — Method

Electrloyte concentration in negative electrode

CaseStudy.SenthilModel.c₂ₚ — Method

Electrloyte concentration in positive electrode

CaseStudy.SenthilModel.c₂ₛ — Method

Electrloyte concentration in separator

CaseStudy.SenthilModel.cₛₙ — Method

Surface solid phase concentration in negative electrode

CaseStudy.SenthilModel.cₛₚ — Method

Surface solid phase concentration in positive electrode

CaseStudy.SenthilModel.jₙ — Method

Local surface reaction rate in negative electrode

CaseStudy.SenthilModel.jₙ₀ — Method

Over-potential in negative electrode

CaseStudy.SenthilModel.jₚ — Method

Local surface reaction rate in positive electrode

CaseStudy.SenthilModel.jₚ₀ — Method

Over-potential in positive electrode

CaseStudy.SenthilModel.k₂ — Method

Electrloyte conductivity

CaseStudy.SenthilModel.k₂ₙ — Method

Electrloyte conductivity in negative electrode

CaseStudy.SenthilModel.k₂ₚ — Method

Electrloyte conductivity in positive electrode

CaseStudy.SenthilModel.k₂ₛ — Method

Electrloyte conductivity in separator

CaseStudy.SenthilModel.rk4 — Method

Simultaneous ODE solver based on the Runge-Kutta 4ᵗʰ order method.

CaseStudy.SenthilModel.ϕ₁ — Method

Overall solid potential function

CaseStudy.SenthilModel.ϕ₁ₙ — Method

Solid potential in negative electrode

CaseStudy.SenthilModel.ϕ₁ₚ — Method

Solid potential in positive electrode

CaseStudy.SenthilModel.ϕ₂ — Method

Overall electrolyte potential function

CaseStudy.SenthilModel.ϕ₂ᵢₙ — Method

Interfacial potential at negative electrode

CaseStudy.SenthilModel.ϕ₂ᵢₚ — Method

Interfacial potential at positive electrode

CaseStudy.SenthilModel.ϕ₂ₙ — Method

Electrloyte potential in negative electrode

CaseStudy.SenthilModel.ϕ₂ₚ — Method

Electrloyte potential in positive electrode

CaseStudy.SenthilModel.ϕ₂ₛ — Method

Electrloyte potential in separator

Constants

CaseStudy.SenthilModel.D₁ₙ — Constant

Solid diffusivity in negative electrode

CaseStudy.SenthilModel.D₁ₚ — Constant

Solid diffusivity in positive electrode

CaseStudy.SenthilModel.D₂ — Constant

Electrolyte diffusivity of material

CaseStudy.SenthilModel.D₂ₙ — Constant

Effective electrolyte diffusivity in the negative electrode

CaseStudy.SenthilModel.D₂ₚ — Constant

Effective electrolyte diffusivity in the positive electrode

CaseStudy.SenthilModel.D₂ₛ — Constant

Effective electrolyte diffusivity in the separator

CaseStudy.SenthilModel.F — Constant

Faraday's constant

CaseStudy.SenthilModel.L — Constant

Total cell thickness

CaseStudy.SenthilModel.R — Constant

Universal gas constant

CaseStudy.SenthilModel.T — Constant

System Temperature (in K)

CaseStudy.SenthilModel.aₙ — Constant

Specific surface area of active material in negative electrode

CaseStudy.SenthilModel.aₚ — Constant

Specific surface area of active material in positive electrode

CaseStudy.SenthilModel.brug — Constant

Bruggeman factor

CaseStudy.SenthilModel.c₁ₙ₀ — Constant

Initial solid phase concentrations in negative electrode

CaseStudy.SenthilModel.c₁ₙₘₐₓ — Constant

Maximum solid phase concentrations in negative electrode

CaseStudy.SenthilModel.c₁ₚ₀ — Constant

Initial solid phase concentrations in positive electrode

CaseStudy.SenthilModel.c₁ₚₘₐₓ — Constant

Maximum solid phase concentrations in positive electrode

CaseStudy.SenthilModel.c₂₀ — Constant

Initial electrolyte concentration

CaseStudy.SenthilModel.kₙ — Constant

Surface reaction rate constants in negative electrode

CaseStudy.SenthilModel.kₚ — Constant

Surface reaction rate constants in positive electrode

CaseStudy.SenthilModel.lₙ — Constant

Thickness of negative electrode

CaseStudy.SenthilModel.lₚ — Constant

Thickness of positive electrode

CaseStudy.SenthilModel.lₛ — Constant

Thickness of negative separator

CaseStudy.SenthilModel.rₙ — Constant

Radii of active material spheres in negative electrode

CaseStudy.SenthilModel.rₚ — Constant

Radii of active material sphere in positive electrode

CaseStudy.SenthilModel.t₊ — Constant

Electrloyte transference number

CaseStudy.SenthilModel.αᵢₙ — Constant

Parameter constant at separator interface of the negative electrode

CaseStudy.SenthilModel.αᵢₚ — Constant

Parameter constant at separator interface of the positive electrode

CaseStudy.SenthilModel.ϵ₁ₙ — Constant

Active material fraction in negative electrode

CaseStudy.SenthilModel.ϵ₁ₚ — Constant

Active material fraction in positive electrode

CaseStudy.SenthilModel.ϵ₁ₛ — Constant

Active material fraction in separator

CaseStudy.SenthilModel.ϵ₂ₙ — Constant

Electrloyte volume fraction in negative electrode

CaseStudy.SenthilModel.ϵ₂ₚ — Constant

Electrloyte volume fraction in positive electrode

CaseStudy.SenthilModel.ϵ₂ₛ — Constant

Electrloyte volume fraction in separator

CaseStudy.AshwiniModel

Modules

CaseStudy.AshwiniModel — Module

Implementation of a closed form reduced order electrochemical model for Lithium-Ion Cells.

For model details refer here.

Types

Functions

Constants