Core

Tip

We suggest that you read through this section if you wish to see how the the library works. To change the model constants (in case you have a different system), see the respective src/{module}/constants/constants.jl file.

LiionBatteryModels

Modules

LiionBatteryModels.LiionBatteryModels — Module

LiionBatteryModels core, allows access to the following submodules:

SenthilModel

AshwiniModel

Types

Functions

Constants

LiionBatteryModels.SenthilModel

Modules

LiionBatteryModels.SenthilModel — Module

Implementation of reduced order model for a lithium ion cell with uniform reaction rate approximation.

For model details refer here.

Structs

LiionBatteryModels.SenthilModel.SolidConcentration — Type

Volume averaged solid concentration

Fields

c₁ₙ: Solid phase concentration in active material spheres in negative electrode.

c₁ₚ: Solid phase concentration in active material spheres in positive electrode.

LiionBatteryModels.SenthilModel.InterfacialConc — Type

Interfacial electrolyte concentration

Fields

c₂ᵢₙ: Interfacial electrolyte concentration in negative electrode.

c₂ᵢₚ: Interfacial electrolyte concentration in positive electrode.

LiionBatteryModels.SenthilModel.InterfacialFlux — Type

Interfacial flux

Fields

q₂ᵢₙ: Interfacial flux in negative electrode.

q₂ᵢₚ: Interfacial flux in positive electrode.

LiionBatteryModels.SenthilModel.SolidRadialConcentration — Type

Radially averaged solid concentration

Fields

c₁ₙᵣ: Solid phase concentration in active material spheres in negative electrode.

c₁ₚᵣ: Solid phase concentration in active material spheres in positive electrode.

Functions

LiionBatteryModels.SenthilModel.SOCₙ — Method

State of Charge in negative electrode

LiionBatteryModels.SenthilModel.SOCₚ — Method

State of Charge in positive electrode

LiionBatteryModels.SenthilModel.Uₙ — Method

Open circuit potential in negative electrode

LiionBatteryModels.SenthilModel.Uₚ — Method

Open circuit potential in positive electrode

LiionBatteryModels.SenthilModel.V — Method

Cell voltage across the battery

LiionBatteryModels.SenthilModel.c₂ — Method

Overall electrolyte concentration function

LiionBatteryModels.SenthilModel.c₂̄ₙ — Method

Average electrolyte concentration in negative electrode

LiionBatteryModels.SenthilModel.c₂̄ₚ — Method

Average electrolyte concentration in positive electrode

LiionBatteryModels.SenthilModel.c₂̄ₛ — Method

Average electrolyte concentration in separator

LiionBatteryModels.SenthilModel.c₂ₙ — Method

Electrloyte concentration in negative electrode

LiionBatteryModels.SenthilModel.c₂ₚ — Method

Electrloyte concentration in positive electrode

LiionBatteryModels.SenthilModel.c₂ₛ — Method

Electrloyte concentration in separator

LiionBatteryModels.SenthilModel.cₛₙ — Method

Surface solid phase concentration in negative electrode

LiionBatteryModels.SenthilModel.cₛₚ — Method

Surface solid phase concentration in positive electrode

LiionBatteryModels.SenthilModel.jₙ — Method

Local surface reaction rate in negative electrode

LiionBatteryModels.SenthilModel.jₙ₀ — Method

Over-potential in negative electrode

LiionBatteryModels.SenthilModel.jₚ — Method

Local surface reaction rate in positive electrode

LiionBatteryModels.SenthilModel.jₚ₀ — Method

Over-potential in positive electrode

LiionBatteryModels.SenthilModel.k₂ — Method

Electrloyte conductivity

LiionBatteryModels.SenthilModel.k₂ₙ — Method

Electrloyte conductivity in negative electrode

LiionBatteryModels.SenthilModel.k₂ₚ — Method

Electrloyte conductivity in positive electrode

LiionBatteryModels.SenthilModel.k₂ₛ — Method

Electrloyte conductivity in separator

LiionBatteryModels.SenthilModel.rk4 — Method

Simultaneous ODE solver based on the Runge-Kutta 4ᵗʰ order method.

LiionBatteryModels.SenthilModel.ϕ₁ — Method

Overall solid potential function

LiionBatteryModels.SenthilModel.ϕ₁ₙ — Method

Solid potential in negative electrode

LiionBatteryModels.SenthilModel.ϕ₁ₚ — Method

Solid potential in positive electrode

LiionBatteryModels.SenthilModel.ϕ₂ — Method

Overall electrolyte potential function

LiionBatteryModels.SenthilModel.ϕ₂ᵢₙ — Method

Interfacial potential at negative electrode

LiionBatteryModels.SenthilModel.ϕ₂ᵢₚ — Method

Interfacial potential at positive electrode

LiionBatteryModels.SenthilModel.ϕ₂ₙ — Method

Electrloyte potential in negative electrode

LiionBatteryModels.SenthilModel.ϕ₂ₚ — Method

Electrloyte potential in positive electrode

LiionBatteryModels.SenthilModel.ϕ₂ₛ — Method

Electrloyte potential in separator

Constants

LiionBatteryModels.SenthilModel.D₁ₙ — Constant

Solid diffusivity in negative electrode

LiionBatteryModels.SenthilModel.D₁ₚ — Constant

Solid diffusivity in positive electrode

LiionBatteryModels.SenthilModel.D₂ — Constant

Electrolyte diffusivity of material

LiionBatteryModels.SenthilModel.D₂ₙ — Constant

Effective electrolyte diffusivity in the negative electrode

LiionBatteryModels.SenthilModel.D₂ₚ — Constant

Effective electrolyte diffusivity in the positive electrode

LiionBatteryModels.SenthilModel.D₂ₛ — Constant

Effective electrolyte diffusivity in the separator

LiionBatteryModels.SenthilModel.F — Constant

Faraday's constant

LiionBatteryModels.SenthilModel.L — Constant

Total cell thickness

LiionBatteryModels.SenthilModel.R — Constant

Universal gas constant

LiionBatteryModels.SenthilModel.T — Constant

System Temperature (in K)

LiionBatteryModels.SenthilModel.aₙ — Constant

Specific surface area of active material in negative electrode

LiionBatteryModels.SenthilModel.aₚ — Constant

Specific surface area of active material in positive electrode

LiionBatteryModels.SenthilModel.brug — Constant

Bruggeman factor

LiionBatteryModels.SenthilModel.c₁ₙ₀ — Constant

Initial solid phase concentrations in negative electrode

LiionBatteryModels.SenthilModel.c₁ₙₘₐₓ — Constant

Maximum solid phase concentrations in negative electrode

LiionBatteryModels.SenthilModel.c₁ₚ₀ — Constant

Initial solid phase concentrations in positive electrode

LiionBatteryModels.SenthilModel.c₁ₚₘₐₓ — Constant

Maximum solid phase concentrations in positive electrode

LiionBatteryModels.SenthilModel.c₂₀ — Constant

Initial electrolyte concentration

LiionBatteryModels.SenthilModel.kₙ — Constant

Surface reaction rate constants in negative electrode

LiionBatteryModels.SenthilModel.kₚ — Constant

Surface reaction rate constants in positive electrode

LiionBatteryModels.SenthilModel.lₙ — Constant

Thickness of negative electrode

LiionBatteryModels.SenthilModel.lₚ — Constant

Thickness of positive electrode

LiionBatteryModels.SenthilModel.lₛ — Constant

Thickness of negative separator

LiionBatteryModels.SenthilModel.rₙ — Constant

Radii of active material spheres in negative electrode

LiionBatteryModels.SenthilModel.rₚ — Constant

Radii of active material sphere in positive electrode

LiionBatteryModels.SenthilModel.t₊ — Constant

Electrloyte transference number

LiionBatteryModels.SenthilModel.αᵢₙ — Constant

Parameter constant at separator interface of the negative electrode

LiionBatteryModels.SenthilModel.αᵢₚ — Constant

Parameter constant at separator interface of the positive electrode

LiionBatteryModels.SenthilModel.ϵ₁ₙ — Constant

Active material fraction in negative electrode

LiionBatteryModels.SenthilModel.ϵ₁ₚ — Constant

Active material fraction in positive electrode

LiionBatteryModels.SenthilModel.ϵ₁ₛ — Constant

Active material fraction in separator

LiionBatteryModels.SenthilModel.ϵ₂ₙ — Constant

Electrloyte volume fraction in negative electrode

LiionBatteryModels.SenthilModel.ϵ₂ₚ — Constant

Electrloyte volume fraction in positive electrode

LiionBatteryModels.SenthilModel.ϵ₂ₛ — Constant

Electrloyte volume fraction in separator

LiionBatteryModels.AshwiniModel

Modules

LiionBatteryModels.AshwiniModel — Module

Implementation of a closed form reduced order electrochemical model for Lithium-Ion Cells.

For model details refer here.

Types

Functions

LiionBatteryModels.AshwiniModel.SOCₙ — Method

State of Charge in negative electrode

LiionBatteryModels.AshwiniModel.SOCₚ — Method

State of Charge in positive electrode

LiionBatteryModels.AshwiniModel.Uₙ — Method

Open circuit potential in negative electrode

LiionBatteryModels.AshwiniModel.Uₚ — Method

Open circuit potential in positive electrode

LiionBatteryModels.AshwiniModel.V — Method

Cell voltage across the battery

LiionBatteryModels.AshwiniModel.c₂ — Method

Overall electrolyte concentration function

LiionBatteryModels.AshwiniModel.c₂̄ₙ — Method

Average electrolyte concentration in negative electrode

LiionBatteryModels.AshwiniModel.c₂̄ₚ — Method

Average electrolyte concentration in positive electrode

LiionBatteryModels.AshwiniModel.c₂̄ₛ — Method

Average electrolyte concentration in separator

LiionBatteryModels.AshwiniModel.c₂ᵢₙ — Method

Interfacial electrolyte concentration in negative electrode

LiionBatteryModels.AshwiniModel.c₂ᵢₚ — Method

Interfacial electrolyte concentration in positive electrode

LiionBatteryModels.AshwiniModel.c₂ₗ — Method

Electrolyte concentration @ x = L in positive electrode.

LiionBatteryModels.AshwiniModel.c₂ₘ — Method

Electrolyte concentration in the middle of battery

LiionBatteryModels.AshwiniModel.c₂ₙ — Method

Electrloyte concentration in negative electrode

LiionBatteryModels.AshwiniModel.c₂ₙ₀ — Method

Electrolyte concentration @ x=0 in negative electrode

LiionBatteryModels.AshwiniModel.c₂ₚ — Method

Electrloyte concentration in positive electrode

LiionBatteryModels.AshwiniModel.c₂ₛ — Method

Electrloyte concentration in separator

LiionBatteryModels.AshwiniModel.cₛₙ — Method

Surface solid phase concentration in negative electrode

LiionBatteryModels.AshwiniModel.cₛₚ — Method

Surface solid phase concentration in positive electrode

LiionBatteryModels.AshwiniModel.d₂ — Method

Electrolyte diffusivity of material

LiionBatteryModels.AshwiniModel.d₂ₙ — Method

Electrolyte diffusivity in negative electrode

LiionBatteryModels.AshwiniModel.d₂ₚ — Method

Electrolyte diffusivity in positive electrode

LiionBatteryModels.AshwiniModel.d₂ₛ — Method

Electrolyte diffusivity in separator

LiionBatteryModels.AshwiniModel.j̅ₙ — Method

Average surface reaction rate in negative electrode

LiionBatteryModels.AshwiniModel.j̅ₚ — Method

Average surface reaction rate in positive electrode

LiionBatteryModels.AshwiniModel.jₙ — Method

Local surface reaction rate in negative electrode

LiionBatteryModels.AshwiniModel.jₙ₀ — Method

Over-potential in negative electrode

LiionBatteryModels.AshwiniModel.jₚ — Method

Local surface reaction rate in positive electrode

LiionBatteryModels.AshwiniModel.jₚ₀ — Method

Over-potential in positive electrode

LiionBatteryModels.AshwiniModel.k₂ — Method

Electrloyte conductivity

LiionBatteryModels.AshwiniModel.k₂ₙ — Method

Electrloyte conductivity in negative electrode

LiionBatteryModels.AshwiniModel.k₂ₚ — Method

Electrloyte conductivity in positive electrode

LiionBatteryModels.AshwiniModel.k₂ₛ — Method

Electrloyte conductivity in separator

LiionBatteryModels.AshwiniModel.param_solver — Method

param_solver

The key function to get all the parameters required to solve for c₂ at a given timestep in the timeseries.

Returns the 16 parameters: n₀(t), n₁(t), n₂(t), n₃(t), p₀(t), p₁(t), p₂(t), p₃(t), s₀(t), s₁(t), s₂(t), <c₂>ₙ, <c₂>ₚ , <c₂>ₛ, c₂|ₓ₌₀ and c₂|ₓ₌ₗ in this order

Params:

- I: The current at the timestep.
- Δt: The span of the timestep.
- c₂ₙₚ: The previous average concentration in negative electrode
- c₂ₛₚ: The The previous average concentration in separator
- c₂ₚₚ: The previous average concentration in positive electrode
- c₂ₙ₀ₚ: The previous concentration in negative electrode @ x=0
- c₂ₚₗₚ: The previous average concentration in positive electrode @x=L
- jn_0: The reaction rate @x=0
- jp_l: The reaction rate @x=L

LiionBatteryModels.AshwiniModel.rk4 — Method

Simultaneous ODE solver based on the Runge-Kutta 4ᵗʰ order method.

LiionBatteryModels.AshwiniModel.ϕ₁ — Method

Overall solid potential function

LiionBatteryModels.AshwiniModel.ϕ₁ₙ — Method

Solid potential in negative electrode

LiionBatteryModels.AshwiniModel.ϕ₁ₚ — Method

Solid potential in positive electrode

LiionBatteryModels.AshwiniModel.ϕ₂ — Method

Overall electrolyte potential function

LiionBatteryModels.AshwiniModel.ϕ₂ᵢₙ — Method

Interfacial potential at negative electrode

LiionBatteryModels.AshwiniModel.ϕ₂ᵢₚ — Method

Interfacial potential at positive electrode

LiionBatteryModels.AshwiniModel.ϕ₂ₙ — Method

Electrloyte potential in negative electrode

LiionBatteryModels.AshwiniModel.ϕ₂ₚ — Method

Electrloyte potential in positive electrode

LiionBatteryModels.AshwiniModel.ϕ₂ₛ — Method

Electrloyte potential in separator

Constants

LiionBatteryModels.AshwiniModel.F — Constant

Faraday's constant

LiionBatteryModels.AshwiniModel.L — Constant

Total cell thickness

LiionBatteryModels.AshwiniModel.R — Constant

Universal gas constant

LiionBatteryModels.AshwiniModel.T — Constant

System Temperature (in K)

LiionBatteryModels.AshwiniModel.aₙ — Constant

Specific surface area of active material in negative electrode

LiionBatteryModels.AshwiniModel.aₚ — Constant

Specific surface area of active material in positive electrode

LiionBatteryModels.AshwiniModel.brugₙ — Constant

Bruggeman factor in negative electrode

LiionBatteryModels.AshwiniModel.brugₚ — Constant

Bruggeman factor in positive electrode

LiionBatteryModels.AshwiniModel.brugₛ — Constant

Bruggeman factor in separator

LiionBatteryModels.AshwiniModel.c₁ₙ₀ — Constant

Initial solid phase concentrations in negative electrode

LiionBatteryModels.AshwiniModel.c₁ₙₘₐₓ — Constant

Maximum solid phase concentrations in negative electrode

LiionBatteryModels.AshwiniModel.c₁ₚ₀ — Constant

Initial solid phase concentrations in positive electrode

LiionBatteryModels.AshwiniModel.c₁ₚₘₐₓ — Constant

Maximum solid phase concentrations in positive electrode

LiionBatteryModels.AshwiniModel.c₂₀ — Constant

Initial electrolyte concentration

LiionBatteryModels.AshwiniModel.k₁ₙ — Constant

Electronic conductivity in negative electode

LiionBatteryModels.AshwiniModel.k₁ₚ — Constant

Electronic conductivity in positive electode

LiionBatteryModels.AshwiniModel.lₙ — Constant

Thickness of negative electrode

LiionBatteryModels.AshwiniModel.lₚ — Constant

Thickness of positive electrode

LiionBatteryModels.AshwiniModel.lₛ — Constant

Thickness of separator

LiionBatteryModels.AshwiniModel.rₙ — Constant

Radii of active material spheres in negative electrode

LiionBatteryModels.AshwiniModel.rₚ — Constant

Radii of active material spheres in positive electrode

LiionBatteryModels.AshwiniModel.t₊ — Constant

Transferennce number of Li ion species dissolved in liquid

LiionBatteryModels.AshwiniModel.αᵪ — Constant

Cathodic transfer coefficient

LiionBatteryModels.AshwiniModel.αₐ — Constant

Anodic transfer coefficient

LiionBatteryModels.AshwiniModel.σ₁ₙ — Constant

Effective electronic conductivity in negative electode

LiionBatteryModels.AshwiniModel.σ₁ₚ — Constant

Effective electronic conductivity in positive electode

LiionBatteryModels.AshwiniModel.ϵ₁ₙ — Constant

Active material fraction in negative electrode

LiionBatteryModels.AshwiniModel.ϵ₁ₚ — Constant

Active material fraction in positive electrode

LiionBatteryModels.AshwiniModel.ϵ₂ₙ — Constant

Volume fraction in negative electrode

LiionBatteryModels.AshwiniModel.ϵ₂ₚ — Constant

Volume fraction in positive electrode

LiionBatteryModels.AshwiniModel.ϵ₂ₛ — Constant

Volume fraction in separator